remove an extraneous layer of wrapper div
